# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cehta Capital API",
    "description": "Backend para la Plataforma Cehta Capital (FIP CEHTA ESG).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_auth_users_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRoleItem"
                  },
                  "title": "Response List Users Api V1 Auth Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{user_id}/role": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Set User Role",
        "operationId": "set_user_role_api_v1_auth_users__user_id__role_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ordenes-compra": {
      "get": {
        "tags": [
          "ordenes-compra"
        ],
        "summary": "List Ocs",
        "operationId": "list_ocs_api_v1_ordenes_compra_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "empresa_codigo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Empresa Codigo"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OrdenCompraListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ordenes-compra"
        ],
        "summary": "Create Oc",
        "operationId": "create_oc_api_v1_ordenes_compra_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenCompraCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenCompraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ordenes-compra/{oc_id}": {
      "get": {
        "tags": [
          "ordenes-compra"
        ],
        "summary": "Get Oc",
        "operationId": "get_oc_api_v1_ordenes_compra__oc_id__get",
        "parameters": [
          {
            "name": "oc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oc Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenCompraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ordenes-compra/{oc_id}/estado": {
      "patch": {
        "tags": [
          "ordenes-compra"
        ],
        "summary": "Update Estado",
        "operationId": "update_estado_api_v1_ordenes_compra__oc_id__estado_patch",
        "parameters": [
          {
            "name": "oc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oc Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenCompraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EstadoUpdateRequest": {
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "emitida",
              "pagada",
              "anulada",
              "parcial"
            ],
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "estado"
        ],
        "title": "EstadoUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OCDetalleCreate": {
        "properties": {
          "item": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Item"
          },
          "descripcion": {
            "type": "string",
            "minLength": 1,
            "title": "Descripcion"
          },
          "precio_unitario": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Precio Unitario"
          },
          "cantidad": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Cantidad"
          }
        },
        "type": "object",
        "required": [
          "item",
          "descripcion",
          "precio_unitario",
          "cantidad"
        ],
        "title": "OCDetalleCreate"
      },
      "OCDetalleRead": {
        "properties": {
          "detalle_id": {
            "type": "integer",
            "title": "Detalle Id"
          },
          "item": {
            "type": "integer",
            "title": "Item"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion"
          },
          "precio_unitario": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Precio Unitario"
          },
          "cantidad": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cantidad"
          },
          "total_linea": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Linea"
          }
        },
        "type": "object",
        "required": [
          "detalle_id",
          "item",
          "descripcion",
          "precio_unitario",
          "cantidad",
          "total_linea"
        ],
        "title": "OCDetalleRead"
      },
...
      "OrdenCompraCreate": {
        "properties": {
          "numero_oc": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Numero Oc"
          },
          "empresa_codigo": {
            "type": "string",
            "title": "Empresa Codigo"
          },
          "proveedor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proveedor Id"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "validez_dias": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Validez Dias",
            "default": 30
          },
          "moneda": {
            "type": "string",
            "enum": [
              "CLP",
              "UF",
              "USD"
            ],
            "title": "Moneda",
            "default": "CLP"
          },
          "neto": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Neto"
          },
          "forma_pago": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forma Pago"
          },
          "plazo_pago": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plazo Pago"
          },
          "observaciones": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OCDetalleCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "numero_oc",
          "empresa_codigo",
          "fecha_emision",
          "neto",
          "items"
        ],
        "title": "OrdenCompraCreate"
      },
      "OrdenCompraListItem": {
        "properties": {
          "oc_id": {
            "type": "integer",
            "title": "Oc Id"
          },
          "numero_oc": {
            "type": "string",
            "title": "Numero Oc"
          },
          "empresa_codigo": {
            "type": "string",
            "title": "Empresa Codigo"
          },
          "proveedor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proveedor Id"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "moneda": {
            "type": "string",
            "title": "Moneda"
          },
          "neto": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Neto"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "oc_id",
          "numero_oc",
          "empresa_codigo",
          "proveedor_id",
          "fecha_emision",
          "moneda",
          "neto",
          "total",
          "estado",
          "pdf_url"
        ],
        "title": "OrdenCompraListItem"
      },
      "OrdenCompraRead": {
        "properties": {
          "oc_id": {
            "type": "integer",
            "title": "Oc Id"
          },
          "numero_oc": {
            "type": "string",
            "title": "Numero Oc"
          },
          "empresa_codigo": {
            "type": "string",
            "title": "Empresa Codigo"
          },
          "proveedor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proveedor Id"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "validez_dias": {
            "type": "integer",
            "title": "Validez Dias"
          },
          "moneda": {
            "type": "string",
            "title": "Moneda"
          },
          "neto": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Neto"
          },
          "iva": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Iva"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "forma_pago": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forma Pago"
          },
          "plazo_pago": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plazo Pago"
          },
          "observaciones": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OCDetalleRead"
            },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "oc_id",
          "numero_oc",
          "empresa_codigo",
          "proveedor_id",
          "fecha_emision",
          "validez_dias",
          "moneda",
          "neto",
          "iva",
          "total",
          "forma_pago",
          "plazo_pago",
          "observaciones",
          "estado",
          "pdf_url",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "OrdenCompraRead"
      },
...
      "Page_OrdenCompraListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrdenCompraListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[OrdenCompraListItem]"
      },
...
      "SetRoleRequest": {
        "properties": {
          "app_role": {
            "type": "string",
            "title": "App Role"
          }
        },
        "type": "object",
        "required": [
          "app_role"
        ],
        "title": "SetRoleRequest"
      },
      "UserMeResponse": {
        "properties": {
          "sub": {
            "type": "string",
            "title": "Sub"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "app_role": {
            "type": "string",
            "title": "App Role"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "sub",
          "email",
          "app_role",
          "allowed_actions"
        ],
        "title": "UserMeResponse"
      },
      "UserRoleItem": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "app_role": {
            "type": "string",
            "title": "App Role"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "app_role"
        ],
        "title": "UserRoleItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}